bugfix global var
Checking for end instead of arguments.end and assigning end instead of arguments.end creates a global var in this model. And it is a singleton, so the second time you will call slice you can check for IsNull(end) but it will always have a value, and you never know which one.